+++ crates/amaru-bootstrap/src/cardano_node/tvar.rs
@@ -22,18 +22,13 @@
 use std::{
     collections::{BTreeMap, BTreeSet},
     io::{Read, Seek, SeekFrom},
-    iter,
 };
 
 use amaru_kernel::{
     Epoch, EraHistory, GlobalParameters, Hash, HeaderHash, MemoizedTransactionOutput, NetworkName, Point,
     StakeCredential, TransactionInput, cbor, cbor::lazy::LazyDecoder,
 };
-use amaru_ledger::{
-    bootstrap::import_initial_snapshot,
-    epoch_transition::GovernanceActivity,
-    store::{self, Store, TransactionalContext},
-};
+use amaru_ledger::{bootstrap::import_initial_snapshot, store::Store};
 use amaru_observability::info;
 use amaru_progress_bar::ProgressBar;
 
@@ -183,26 +178,7 @@ where
         actual_size += size;
 
         if !utxo.is_empty() {
-            let transaction = db.create_transaction();
-            transaction.save(
-                era_history,
-                &protocol_parameters,
-                GovernanceActivity::default(),
-                point,
-                None,
-                store::Columns {
-                    utxo: utxo.into_iter(),
-                    pools: iter::empty(),
-                    accounts: iter::empty(),
-                    dreps: iter::empty(),
-                    cc_members: iter::empty(),
-                    proposals: iter::empty(),
-                    votes: iter::empty(),
-                },
-                Default::default(),
-                iter::empty(),
-            )?;
-            transaction.commit()?;
+            db.save_bootstrap_utxo(era_history, &protocol_parameters, point, utxo.into_iter())?;
         }
 
         if done {